Fix spelling mistake

+++ installation/sql/postgresql/extensions.sql
@@ -899,7 +899,7 @@ CREATE INDEX "#__scheduler_tasks_idx_checked_out" ON "#__scheduler_tasks" ("chec
 CREATE TABLE IF NOT EXISTS "#__schemaorg"
 (
   "id" serial NOT NULL,
-  "itemId" smallint,
+  "itemId" integer,
   "context" varchar(100),
   "schemaType" varchar(100),
   "schemaForm" text,
